+++ killer.h
@@ -15,12 +15,12 @@ extern "C" {
     int kill_fish(numpl_array * array);
     int kill_xywing(numpl_array * array);
 
-    int64_t analyze_single(numpl_array * array, solve_info * info);
-    int64_t analyze_hidden_single(numpl_array * array, solve_info * info);
-    int64_t analyze_locked_candidate(numpl_array * array, solve_info * info);
-    int64_t analyze_tuple(numpl_array * array, solve_info * info);
-    int64_t analyze_fish(numpl_array * array, solve_info * info);
-    int64_t analyze_xywing(numpl_array * array, solve_info * info);
+    int analyze_single(numpl_array * array, solve_info * info);
+    int analyze_hidden_single(numpl_array * array, solve_info * info);
+    int analyze_locked_candidate(numpl_array * array, solve_info * info);
+    int analyze_tuple(numpl_array * array, solve_info * info);
+    int analyze_fish(numpl_array * array, solve_info * info);
+    int analyze_xywing(numpl_array * array, solve_info * info);
 
 #if defined(__cplusplus)
 }
